requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.16.1
firebase-admin==6.6.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.1.4
pypng==0.20220715.0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.3
tokenizers==0.20.3
torch
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
webdriver-manager==4.0.2
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
